requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.0.11
dnspython==2.0.0
filetype==1.0.5
Flask==1.1.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
iniconfig==1.0.1
instabot==0.117.0
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.5.0
moviepy==0.2.3.2
numpy==1.19.1
packaging==20.4
pika==1.1.0
Pillow==7.2.0
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
proglog==0.1.9
py==1.9.0
pymongo==3.11.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.0
schedule==0.6.0
six==1.15.0
toml==0.10.1
tqdm==4.11.2
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.1.0